Rest API in progress

+++ website/generate_pairs/routes.py
@@ -3,7 +3,7 @@ from flask_login import current_user, login_required
 from website.forms.GenerateRandomPairs import GenerateRandomPairsForm
 from website.database.DB import session
 from website.database.models import RandomPairs, RandomPairsResults
-from website.generate_pairs.generate_random_pairs import generate_random_pairs, RandomPerson
+from website.generate_pairs.generate_random_pairs import RandomPerson, generate_random_pairs
 import json
 
 
@@ -26,15 +26,16 @@ def pairs():
             try:
                 session.add(random_person)
                 session.commit()
-            except Exception:
+            except Exception as err:
                 session.rollback()
-                raise
+                print(err)
             finally:
                 session.close()
 
         return redirect('/generate-pairs')
 
-    user_pair = session.query(RandomPairs).filter_by(user_id=current_user.id).all()
+    user_pair = session.query(
+        RandomPairs).filter_by(user_id=current_user.id).all()
 
     return render_template(
         'generate_pairs.html',
@@ -50,26 +51,30 @@ def delete_pair():
     pair = session.query(
         RandomPairs).filter_by(user_id=current_user.id).order_by(RandomPairs.id.desc()).first()
 
-    with session:
-        try:
-            session.delete(pair)
-            session.commit()
-        except Exception:
-            session.rollback()
-            raise
-        finally:
-            session.close()
+    if pair:
+        with session:
+            try:
+                session.delete(pair)
+                session.commit()
+            except Exception:
+                session.rollback()
+                raise
+            finally:
+                session.close()
 
+        return redirect('/generate-pairs')
+
+    flash('There is no pairs!', 'danger')
     return redirect('/generate-pairs')
 
 
-@generate_pairs.route('/results', methods=['GET', 'POST'])
+@generate_pairs.route('/results', methods=['POST'])
 @login_required
 def results():
-    user_random_person_pool = session.query(RandomPairs).filter_by(user_id=current_user.id).all()
+    user_random_person_pool = session.query(
+        RandomPairs).filter_by(user_id=current_user.id).all()
 
-    if len(user_random_person_pool) % 2 == 0 and len(user_random_person_pool) != 0:
-
+    if len(user_random_person_pool) != 0 and len(user_random_person_pool) > 1:
         random_person_pool = []
 
         for row in user_random_person_pool:
@@ -77,19 +82,9 @@ def results():
                 RandomPerson(row.random_person_name, row.random_person_email)
             )
 
-        def call_generate_random_pairs():
-            random_pairs = generate_random_pairs(random_person_pool)
-            pairs_list = []
-
-            for [person_one, person_two] in random_pairs:
-                pairs_list.append([f'{person_one.name}, {person_one.email}', f'{person_two.name}, {person_two.email}'])
-
-            return json.dumps(pairs_list)
-
-        execute_call_generate_random_pairs = call_generate_random_pairs()
-
+        user_results = generate_random_pairs(random_person_pool)
         user_random_pairs = RandomPairsResults(
-            results=execute_call_generate_random_pairs,
+            results=user_results,
             user_id=current_user.id
         )
 
@@ -98,29 +93,27 @@ def results():
                 session.add(user_random_pairs)
                 session.query(RandomPairs).filter_by(user_id=current_user.id).delete()
                 session.commit()
-            except Exception:
+            except Exception as err:
                 session.rollback()
-                raise
+                print(err)
             finally:
                 session.close()
 
         return redirect('/show-results')
 
-    elif len(user_random_person_pool) == 0:
-        flash('Add your pairs!', 'danger')
-
-        return redirect('/generate-pairs')
-
-    else:
-        flash('The number of pairs has to be equal', 'danger')
-
-        return redirect('/generate-pairs')
+    flash('Add your pairs!', 'danger')
+    return redirect('/generate-pairs')
 
 
 @generate_pairs.route('/show-results')
 @login_required
 def show_results():
-    user_random_pairs_result = session.query(RandomPairsResults).filter_by(user_id=current_user.id).all()
+    user_random_pairs_result = session.query(
+        RandomPairsResults).filter_by(user_id=current_user.id).all()
+    flat_results = [result for result in user_random_pairs_result]
+    print(flat_results)
+    print(user_random_pairs_result)
+    # TODO ogarnąć dostęp do tych list (results)
 
     return render_template(
         'results.html',
@@ -132,18 +125,19 @@ def show_results():
 @login_required
 def delete_result():
     result_to_delete = json.loads(request.data)
-    result_id = result_to_delete.get('resultId')
+    result_id = result_to_delete['result_id']
     result_query = session.query(RandomPairsResults).get(result_id)
 
-    if result_query and result_query.user_id == current_user.id:
-        with session:
-            try:
-                session.delete(result_query)
-                session.commit()
-            except Exception:
-                session.rollback()
-                raise
-            finally:
-                session.close()
+    if result_query:
+        if result_query.user_id == current_user.id:
+            with session:
+                try:
+                    session.delete(result_query)
+                    session.commit()
+                except Exception as err:
+                    session.rollback()
+                    print(err)
+                finally:
+                    session.close()
 
-        return jsonify({})
+            return jsonify({})
